+++ lang/nl/validation.php
@@ -93,10 +93,18 @@ return [
         'plural' => 'Fields :values are not countable.',
     ],
 
+    'boolean_string' => 'Het :attribute moet waar of onwaar zijn.',
+
     /** @TODO requires translation */
     'client_id' => 'The :attribute format is invalid.',
 
     'date_time_iso_8601' => 'Het attribuut :attribute heeft geen geldig ISO 8601 datum/tijd formaat.',
+
+    'json_boolean' => 'Het veld :attribute moet een boolean zijn.',
+
+    'json_integer' => 'Het veld :attribute moet een geheel getal zijn.',
+
+    'json_number' => 'Het veld :attribute moet een nummer zijn.',
 
     'parameter_not_supported' => 'Parameter :name is niet toegestaan.',